Tried Updating State w/ User Interaction
Tried to add the feature of being able to click on a team inside of a game, to advance that team to the next round, but not currently understanding how to properly use useState. During debugging, came to the realization that each individual game does not have its own element. Probably need to solve for that first - might just be a question of adding a "bracket" component to contain the entire thing.