Organizing by Rounds - WIP
Started to go down the road of trying to organize the front end appearance by round, to make it clear to end users. Struggling to find a way to accurately start / end the columns right, based on the iteration happening in the various components. Wondering if a round component is going to be necessary between the bracket and game components that exist today.